🪓 Code separation for utils (🧹clean code)
