+++ internal/utils/assertions_test.go
@@ -1,0 +1,13 @@
+// ⚡️ Fiber is an Express inspired web framework written in Go with ☕️
+// 🤖 Github Repository: https://github.com/gofiber/fiber
+// 📌 API Documentation: https://docs.gofiber.io
+
+package utils
+
+import "testing"
+
+func Test_Utils_AssertEqual(t *testing.T) {
+	t.Parallel()
+	AssertEqual(nil, []string{}, []string{})
+	AssertEqual(t, []string{}, []string{})
+}
